Remove user file check for install.
User defined adapter is checked during the build
process for initramfs, the adapter not existing
before the package is installed should not cause
the installation to fail.

No functional change as this is already checked in
hooks as a pre-requisite to build, and allows
'make install' to succeed with a default package.

Changed:
* build_wg_setconf - remove check for adapter
  configuration existing during installation.